Morningstar webscraper.
I noticed that Morningstar had market data available in minute intervals so I was curious how to get to that information. Though Morningstar offers an API, I wanted to get to the data myself. To display the data one has open a chart widget then enters time spans and coarseness.