DOI Validation Program
I made this program while I was working in library science. This project was abandoned in early 2019 due to a breach of contract with publishers. The input files and supporting database are "mock" datasets and do not reflect the entirety of what we used for this project. This was an early webscraper for me so pauses were not implemented (this caused some problems) and some of the websites may have changed their HTML and Javascript so the program may not be as versatile as it appears. I had known much about XPaths so I used BeautifulSoup to search for elements, something which can be improved upon.